Move NIP-65 semantics to protocol
