+++ web/src/lib/nostr/protocol/nip65.test.ts
