+++ web/src/lib/nostr/protocol/nip65.ts
